+++ instrumentation/opentelemetry-instrumentation-django/tests/test_sqlcommenter.py
@@ -88,7 +88,7 @@ class TestMiddleware(WsgiTestBase):
         execute_mock_obj = MagicMock()
         qw_instance(
             execute_mock_obj,
-            "Select 1",
+            "Select 1;",
             MagicMock("test"),
             MagicMock("test1"),
             MagicMock(),
@@ -97,7 +97,7 @@ class TestMiddleware(WsgiTestBase):
         self.assertEqual(
             output_sql,
             "Select 1 /*app_name='app',controller='view',route='route',traceparent='%%2Atraceparent%%3D%%2700-0000000"
-            "00000000000000000deadbeef-000000000000beef-00'*/",
+            "00000000000000000deadbeef-000000000000beef-00'*/;",
         )
 
     @patch(