+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://github.com/open-telemetry/opentelemetry-python/compare/v1.12.0rc2-0.32b0...HEAD)
 - Adding multiple db connections support for django-instrumentation's sqlcommenter
   ([#1187](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1187))
+- SQLCommenter semicolon bug fix
+  ([#1200](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1200/files))
 
 ### Added
 - `opentelemetry-instrumentation-redis` add support to instrument RedisCluster clients